+++ lib/verquest/properties/array.rb
@@ -21,30 +21,38 @@ module Verquest
       # @param type [String, Symbol] The type of items in the array, can be a default type or a custom field type
       # @param map [String, nil] The mapping path for this property (nil for no explicit mapping)
       # @param required [Boolean] Whether this property is required
+      # @param nullable [Boolean] Whether this property can be null
       # @param item_schema_options [Hash] Additional JSON schema options for the array items (merged with custom type options)
       # @param schema_options [Hash] Additional JSON schema options for the array property itself
       # @raise [ArgumentError] If type is not one of the allowed types (default or custom)
       # @raise [ArgumentError] If attempting to map an array to the root
-      def initialize(name:, type:, map: nil, required: false, item_schema_options: {}, **schema_options)
+      def initialize(name:, type:, map: nil, required: false, nullable: false, item_schema_options: {}, **schema_options)
         raise ArgumentError, "Type must be one of #{allowed_types.join(", ")}" unless allowed_types.include?(type.to_s)
         raise ArgumentError, "You can not map array to the root" if map == "/"
 
         if (custom_type = Verquest.configuration.custom_field_types[type.to_sym])
-          @type = custom_type[:type].to_s
+          @item_type = custom_type[:type].to_s
           @item_schema_options = if custom_type.key?(:schema_options)
             custom_type[:schema_options].merge(item_schema_options).transform_keys(&:to_s)
           else
             item_schema_options.transform_keys(&:to_s)
           end
         else
-          @type = type.to_s
+          @item_type = type.to_s
           @item_schema_options = item_schema_options.transform_keys(&:to_s)
         end
 
         @name = name.to_s
         @map = map
         @required = required
+        @nullable = nullable
         @schema_options = schema_options&.transform_keys(&:to_s)
+
+        @type = if nullable
+          %w[array null]
+        else
+          "array"
+        end
       end
 
       # Generate JSON schema definition for this array property
@@ -53,8 +61,8 @@ module Verquest
       def to_schema
         {
           name => {
-            "type" => "array",
-            "items" => {"type" => type}.merge(item_schema_options)
+            "type" => type,
+            "items" => {"type" => item_type}.merge(item_schema_options)
           }.merge(schema_options)
         }
       end
@@ -72,7 +80,7 @@ module Verquest
 
       private
 
-      attr_reader :type, :schema_options, :item_schema_options
+      attr_reader :type, :item_type, :schema_options, :item_schema_options
 
       # Gets the list of allowed item types, including both default and custom types
       #